Move LoadingButton out of index
Test prep
